pyproject.toml
[project]
name = "sam3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albucore==0.0.24",
    "albumentations==2.0.8",
    "annotated-types==0.7.0",
    "cachetools==6.2.2",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "colorama==0.4.6",
    "decord>=0.6.0",
    "einops>=0.8.2",
    "geojson==3.2.0",
    "google-api-core==2.28.1",
    "google-auth==2.43.0",
    "google-cloud-vision==3.11.0",
    "googleapis-common-protos==1.72.0",
    "gradio>=6.6.0",
    "grpcio==1.76.0",
    "grpcio-status==1.76.0",
    "huggingface-hub>=1.4.1",
    "idna==3.11",
    "iopath>=0.1.10",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "numpy==2.3.5",
    "opencv-contrib-python==4.11.0.86",
    "opencv-python==4.11.0.86",
    "opencv-python-headless==4.11.0.86",
    "pillow==12.0.0",
    "proto-plus==1.26.1",
    "protobuf==6.33.2",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycocotools==2.0.10",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pyyaml==6.0.3",
    "rasterio>=1.5.0",
    "requests==2.32.5",
    "rsa==4.9.1",
    "scikit-image>=0.26.0",
    "scipy==1.16.3",
    "segment-anything-model>=0.0.1.dev0",
    "segment-geospatial[samgeo3]>=1.2.2",
    "setuptools<82",
    "shapely==2.1.2",
    "simsimd==6.5.3",
    "stringzilla==4.4.0",
    "timm>=1.0.24",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "ultralytics>=8.3.237",
    "urllib3==2.6.1",
    "whitebox>=2.3.6",
]
